server/requirements.txt
flask
python-dotenv
flask-cors
google-generativeai
langfuse